// types.ts
export interface ReadingQuestion {
  question: string;
  options: {
    key: string;
    value: string;
  }[];
  correctAnswer: string;
}

export interface KeyVocabulary {
  word: string;
  meaning: string;
}

export interface ReadingAnalysisResult {
  summary: string;
  vocabulary: KeyVocabulary[];
  questions: ReadingQuestion[];
}